Resolved dynamic rendering check fail
