+++ app/api/payment/route.ts
@@ -1,6 +1,8 @@
 import { NextResponse } from "next/server";
 import { createClient } from "@/utils/supabase/server";
 
+export const dynamic = "force-dynamic";
+
 let Stripe: any = null;
 let stripe: any = null;
