+++ app/api/paypal/verify/route.ts
@@ -1,3 +1,5 @@
+export const dynamic = "force-dynamic";
+
 export async function POST(request: Request) {
   try {
     const body = await request.json();
